Update BT service. Fix signal fragments. Update other activities.

+++ app/src/main/java/com/example/helio/arduino/core/Constants.java
@@ -2,10 +2,10 @@ package com.example.helio.arduino.core;
 
 public class Constants {
 
-    public static final int HZ = 1;
-    public static final int kHZ = HZ * 1000;
-    public static final int MHZ = kHZ * 1000;
-    public static final int MAX_HZ = MHZ * 20;
+    public static final long HZ = 1;
+    public static final long kHZ = HZ * 1000;
+    public static final long MHZ = kHZ * 1000;
+    public static final long MAX_HZ = MHZ * 20;
 
     // Message types sent from the BluetoothChatService Handler
     public static final int MESSAGE_STATE_CHANGE = 1;
